% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_gender.R
\name{assign_gender}
\alias{assign_gender}
\title{Infer people's gender}
\usage{
assign_gender(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Looks up for the gender of the personal names in \code{x}.
}
\details{
Unidentified genders are set as \code{NA} values.
}
\examples{
x. = c("Carlos", "Martha", "not_a_name")
assign_gender(x.)


library(tidyverse)

edcn_genders <- edcn \%>\%
 separate(subject_name,
          into = c("family_names", "personal_name"),
          sep = ",") \%>\%
 mutate(gender = assign_gender(personal_name)) \%>\%
 select(1:2, gender, everything())
}
