% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{atlas}
\alias{atlas}
\title{Find Spanish towns on a map}
\usage{
atlas(x, division = "provincia")
}
\arguments{
\item{x}{A character or an integer vector with the names or the codes of the
Spanish divisions to map.}

\item{division}{The type of territorial division in \code{x}: "municipio",
"provincia" or "ccaa".}
}
\value{
A HTML widget object, see \link[leaflet]{leaflet}.
}
\description{
Renders a map plotting all the towns of \code{x}. Their
official names and a list with further information can be
accessed by hovering and clicking on the markers, respectively.
}
\examples{
atlas("Navarra")
}
