% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias}
\alias{bias}
\title{Set bias parameters}
\usage{
bias(division = "provincia", towards = NULL, against = NULL)
}
\arguments{
\item{division}{Biased division type: "municipio, "provincia" or "ccaa".}

\item{towards}{Codes of the divisions to bias towards: integer vector.}

\item{against}{Codes of the divisions to bias against: integer vector.}
}
\value{
A three-element list with class \code{bias}.
}
\description{
Bias parameters define the action scope of \link{provinciator} by favoring
its string matching process towards or against selected regional divisions.
}
\details{
Use \link{codigator} to retrieve the requested codes or execute
  \link{codex} to print the full list.
}
