% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age}
\alias{calculate_age}
\title{Determine people's age}
\usage{
calculate_age(givendate, birthdate)
}
\arguments{
\item{givendate, birthdate}{POSIXt, Date or character vectors of equal length.}
}
\value{
An integer vector of the same length as \code{givendate}.
}
\description{
Finds out the calendar year difference between two dates.
}
\details{
For \code{birthdate > givendate} \code{NA}s are returned .
}
\examples{
calculate_age(Sys.Date(), "1988-02-29")
}
