% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_name.R
\name{clean_name}
\alias{clean_name}
\title{Homogenize personal names' syntax}
\usage{
clean_name(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Removes impossible punctuation characters and sets the proper capitalization
and spacing of the names in \code{x}.
}
\examples{
x. <- c("Alberto De La Fue+nte Y Pastor", " JULIA O´DONNELL ?")
clean_name(x.)
}
