% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex.R
\name{codex}
\alias{codex}
\title{Print all main Spanish territorial division codes}
\usage{
codex(order_by = "provincia", ascending = TRUE)
}
\arguments{
\item{order_by}{Arranges output either by "provincia" or "ccaa".}

\item{ascending}{Logical for whether to use ascending order or not.}
}
\value{
A tibble with all the "provincias" and "comunidades autónomas"
  along with their codes.
}
\description{
Prints the official codes of all the Spanish "provincias" and
  "comunidades autónomas".
}
\examples{
codex()
}
