% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigator.R
\name{codigator}
\alias{codigator}
\title{Find Spanish territorial division codes}
\usage{
codigator(x, division = "provincia")
}
\arguments{
\item{x}{A character or numeric vector (see \emph{Value}).}

\item{division}{The type of territorial division to look for:
\code{"municipio"}, \code{"provincia"} or \code{"ccaa"}.}
}
\value{
A tibble with \code{x} and the corresponding matches and codes. If
  \code{x} is numeric the output will be the name of the matched territorial
  divisions.
}
\description{
Provides the
  \href{https://www.ine.es/daco/daco42/codmun/codmunmapa.htm}{INE}'s official
  code of the Spanish territorial divisions in x. It uses approximate string
  matching.
}
\examples{
x <- c("Pontevedra", "Madrid", "Teruel", "Valencia")
codigator(x)

}
