% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provinciator.R
\name{provinciator}
\alias{provinciator}
\alias{provinciate}
\title{Locate and homogenize Spanish territorial divisions}
\usage{
provinciator(x, division = "provincia", scope = bias())

provinciate(.data, .var, ...)
}
\arguments{
\item{x}{A character vector.}

\item{division}{The type of territorial division in \code{x}:
\code{"municipio"} (includes "unidades poblacionales"), \code{"provincia"}
or \code{"ccaa"}.}

\item{scope}{A list with class \link{bias} used to favor the string matching
process of \code{x} towards or against selected regional divisions.}

\item{.data}{A data frame.}

\item{.var}{A variable from \code{.data}.}

\item{...}{Other arguments passed to \code{provinciator} (except \code{x}).}
}
\value{
For \code{provinciator}, a tibble (i.e. \link[tibble]{tbl_df}) with
 the \href{https://www.ine.es/daco/daco42/codmun/codmunmapa.htm}{INE}'s
 official division names in \code{x} plus their geographical coordinates.

 For \code{provinciate}, an object of the same class as \code{.data} with the
 results attached as new columns.
}
\description{
Uses approximate string matching to infer the entity names in
 \code{x} and returns their official designation and geographical
 coordinates.
}
\details{
Geographical coordinates for \code{"provincia"} and \code{"ccaa"}
 refer to the capital of the chosen division type.

 If \code{scope} is used, follow this example as a guide:

 \code{scope = bias("provincia", c(1, 20, 48:49), 4)}

 Basque provinces (1, 20 & 48) plus Zamora (49) will be prioritized when
 matching names in \code{x} and Almería (4) won't be used at all.
}
\note{
By default, tibbles print no more than 3 decimals. Use
 \code{options(pillar.sigfig = n)} to change this number.
}
\examples{
x. <- c("Almería", "Teruel", "Lérida")
provinciator(x.)


library(tidyverse)

edcn_towns <- edcn \%>\%
  provinciate(town_of_birth, "municipio") \%>\%
  select(town_of_birth, last_col(2:0), everything())

}
\seealso{
Use \link{atlas} to learn about the town names used by
 \code{provinciator}.
}
