% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_extra_spaces.R
\name{remove_extra_spaces}
\alias{remove_extra_spaces}
\title{Trim extra white-spaces from strings}
\usage{
remove_extra_spaces(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
Removes unnecessary white-spaces from the left, right and/or interior of
strings.
}
\examples{
x. <- c(" San  Sebastian ")
remove_extra_spaces(x.)
}
