% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_vowel_diacrtics.R
\name{remove_vowel_diacritics}
\alias{remove_vowel_diacritics}
\title{Get rid of vowel diacritics}
\usage{
remove_vowel_diacritics(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
Removes vowel diacritics without changing the underlying encoding.
}
\examples{
x. <- "{Múñez} Muñana; Ávila; Castilla y León"
remove_vowel_diacritics(x.)
}
